Fix #3: Make tests run more quickly.
Add a new shell script to run the tests.  It uses a new new environment
variable `HREPL_TEST_OUTPUT` that lets the tests share the same output
directory.  Run `bazel clean` in between each test to make them essentially
hermetic (as far as this test is concerned), without redoing expensive
computations in each test such as fetching/installing the GHC build or building
the `protoc` binary.

Also set the disk cache explicitly in the test output directory to cache
build outputs across tests.

When I ran the tests with With this change, when I ran it two tests each
took a little under 5 minutes (`joined_ffi_test`, because it was the first
one, and `proto_test`, because it needed to rebuild more protobuf dependencies)
and the rest of the tests each took ~10s.  The total runtime was a little
under 10 minutes.