Run ghci in the current directory, rather than the execroot.
After this change, it may be run from any subdirectory of the
Bazel workspace.

Note; this is not perfect, sice some Templatehaskell code may
expect to run from a specific root (e.g. with `file-embed`).
However, this is a more general issue with GHC, and stack
(for example) has the same issue:
https://github.com/commercialhaskell/stack/issues/1057