Filter for haskell_binary|library|test
These are the only kinds of targets that produce the
`compile_info_output_groups` that `hrepl` requires. This is to avoid
errors if non-Haskell targets are picked up as intermediate dependencies
between targets specified on the command-line, e.g. through `data`
dependencies.

Additionally, this enables users to call `hrepl` on wild-card targets.
E.g. in the `rules_haskell` repository
```
hrepl //tests/binary-with-lib:all
```
Now no longer fails on with an error of the form
```
hrepl: .../tests/binary-with-lib/binary-with-lib@repl.HaskellCompile.pb: openBinaryFile: does not exist (No such file or directory)
```